CommandIntegrator version 1.2.1:
New closure, enabling developers to implicitly allow their methods to
accept a keyword argument "channel", which is useful when calling them
from an infinity loop such as in a scheduling system or alike, where
not all messages should show in the same channel/pane/site in the front
end. Using this decorator above the definition of a method allows the
method to accept the argument 'channel' which will result in a dictionary
returned with the method return under 'result' key and the channel that
was entered from the caller, under the 'channel' key.