the __getitem__ method is written to return None if the callback does not match the given command, otherwise return the function associated with it (self._func)
The callback instance will evaluate whether it contains any of the
words provided in the 'lead' property first.
If it does, and 'trail' property is left undefined, callback returns
the function.

If the 'trail' property is defined however - the message must also
comply with any of the words in the trail parameter (string or collection
of strings) - and all of these must occur _after_ the 'lead' strings
to comply.

If the 'ordered' property is set to True, in addition to the rules above,
the message must _also_ comply with having every word in both
lead and trail, if trail is defined, in the same order of appearance as
in 'lead' and(if) _trail_.
This is offloaded to a separate method, _assert_ordered.