Bug fix: #6
using deepcopy with the discord.Message object causes an unforseen
crash where serialization of the object with pickle by deepcopy causes
the method call to crash, which leads to a halt in the async job.

This is fixed by moving the only temporary stripping of characters for
identifying a callback bind to the method "matches" in the class
Callback itself.

It resolves the previous bug (5) as well as this resulting bug.