the Schedule tool bundle
The Schedule tool bundle in the framework to allow for pythonic
and semantically pleasing scheduling of callables within an application.

The Schedule is designed for ease of use with sensible parameter naming
and support for threaded jobs to allow for either an async or
multi threaded development approach, prohibiting scheduled functions
to block and crash any other calling code in the stack.

The Schedule module can be used either with the decorator @schedule
or by adding to the static object "Schedule" by using the Schedule.run()
method, and providing a pointer to the function for scheduling with
provided arguments for when the function should be called.

Scheduling can be configured to run a callable after a set period
of time, at a certain time, have it repeat a certain weekday or
a selection of week days, every hour, day, minute or second
and is very flexible with combinations of these settings to allow
for a very free way to schedule functions.